Rename azure/compute/os to azure/compute/azureos
